Refactor from assetIds to token addresses with an icon lookup instead
