+++ apps/fixed-borrow/src/ui/token/AssetBadge.tsx
@@ -1,16 +1,17 @@
 import { ReactElement } from "react";
 import { AssetIcon } from "src/ui/token/AssetIcon";
-import { AssetId } from "src/ui/token/AssetId";
+import { Address, useToken } from "wagmi";
 
 interface AssetBadgeProps {
-  assetId: AssetId;
+  address: Address;
 }
 
-export function AssetBadge({ assetId }: AssetBadgeProps): ReactElement {
+export function AssetBadge({ address }: AssetBadgeProps): ReactElement {
+  const { data: tokenMetadata } = useToken({ address });
   return (
     <div className="flex h-8 items-center gap-1.5 rounded-full bg-dawn pl-1 pr-2">
-      <AssetIcon assetId={assetId} />
-      <span className="text-body">{assetId}</span>
+      <AssetIcon address={address} />
+      <span className="text-body">{tokenMetadata?.symbol}</span>
     </div>
   );
 }
