--- apps/fixed-borrow/src/ui/token/AssetId.ts
@@ -1,1 +1,0 @@
-export type AssetId = "DAI" | "wstETH";
